% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inference}
\alias{inference}
\title{Results obtained running VERSO on the provided input dataset.}
\format{
results obtained running VERSO on the provided input dataset
}
\usage{
data(inference)
}
\value{
results obtained running VERSO on the provided input dataset
}
\description{
Results obtained running VERSO on the provided input dataset.
}
