% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variants}
\alias{variants}
\title{Mutation data obtained by variant calling from raw data of a 
selected set of SARS-CoV-2 samples available from NCBI 
BioProject PRJNA610428.}
\format{
SARS-CoV-2 variants
}
\source{
NCBI BioProject PRJNA610428
}
\usage{
data(variants)
}
\value{
SARS-CoV-2 variants
}
\description{
The dataset includes variants for a selected set of 15 
SARS-CoV-2 samples obtained by variant calling from raw data 
available from NCBI BioProject PRJNA610428.
}
