% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend.R
\name{write.newick.tree}
\alias{write.newick.tree}
\title{write.newick.tree}
\usage{
write.newick.tree(phylogenetic_tree, phylogeny_file = "phylogenetic_tree.new")
}
\arguments{
\item{phylogenetic_tree}{Inference results by VERSO.}

\item{phylogeny_file}{File where to save the phylogenetic tree in newick format.}
}
\description{
Write a phylogenetic tree as inferred by VERSO to a newick format file.
}
\examples{
data(inference)
write.newick.tree(phylogenetic_tree = inference, 
                  phylogeny_file = "inference_tree.new")

}
